--- frontend/src/app/(user)/components/Chart/PieChart.stories.tsx
@@ -1,17 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-
-import PieChart from './PieChart';
-
-const meta: Meta<typeof PieChart> = {
-    component: PieChart,
-    title: 'atoms/PieChart',
-    tags: ['autodocs'],
-    argTypes: {}
-};
-export default meta;
-
-type Story = StoryObj<typeof PieChart>;
-
-export const Default: Story = {
-    args: {}
-};
